Create initial tests for parallel time_delay_estimator.
With the parallel (aka batched) FFT transforms implemented
we can start the work to implement the time_delay_estimator.
These changes just make sure the right code compiles.
The plan is to write the tests and code more or less in the TDD
style, initially all the code goes in the unit test file, I
will move it to its own header later.