Make sure jb::opencl::copy_to_host_async() waits for the previous events.
I wrote a unit test for jb::opencl::copy_to_host_async(), mostly to
improve code coverage, but in the process found that I had forgotten
to pass the wait set to the enqueue_read_buffer_async() call.
The test was able to reproduce the problem, and it is now fixed.