First draft of the jb::fftw::time_delay_estimator_many.
This is WIP to define the interface of a time_delay_estimator,
the current version offers a blocking interface, while we
know we need to work towards an asynchronous one.  I just
want to have a basic implementation to start writing tests
and benchmarks.
Some minimal amount of refactoring done here too.