Document power analysis and many changes on the analysis output.
First we document the desired effects to measure, and the number
of samples required to achieve the desired statistical power.
Then we document different ways to run the analysis, none of them
are properly scripted, but they are close enough to serve as
reminders when running this as a notebook.

Other than the power analysis this is looking more like a series of
notes than a proper script.