Optimized jb::itch5::price_levels() and reduce calls to it.
Callgrind claims that this is one of the most expensive functions
in the code, I have reduced the estimated cost from 222 million
cycles to 212 million cycles (in the benchmark, with 10,000,000
calls).  I also did some cleanup to reduce the number of calls to
it.
The next most expensive function in the array book class is
move_to_bottom(), maybe there are some opportunities there.