Including package versions in metadata.
Make more of an effort to capture the shared library
versions.  Capture their path, figure out what package
provides them, and then print out the package version.
Also capture the version of the libstdc++ library.
And I was including the kernel version used to build the program,
but not the version used to run it, so add that too.
Some minor fixes to the output when setting the RT scheduling limits.