Fix failures in test due to excessive numerical zeal.
The test was suppose to valid the basic flow and arithmetic
of the function, but it devolved (by accident) into a test
of the inherent errors in floating point arithmetic.
I changed the inputs to make the failures due to rounding go away.
Also add pragma about floating point to make pocl happy.