First benchmark running in container.
Many improvements in the generate_common.sh shell library
to gracefully degrade when the environment does not contain
the necessary tools.  That way the images can be smaller and
we still get results, but maybe not as much detail about the
environment as we would get otherwise.
Also more fixes around searching for binaries for both container
and development runs of the benchmarks.
Because the source code is not available in the runtime containers
I created a program to capture the build-time information, such
as compiler versions, flags, and git revisions.
Avoid bloat in the runtime image by *not* updating the packages,
just update the metadata cache, and download strictly what we
need (and even a little less :-)