Refactor generation of Doxygen documentation.
The script to generate Doxygen documentation was partially in
ci/gendocs.sh and partially in the .travis.yml file.
Clean all of that up into two scripts, one that runs inside
the a docker container to create the Doxygen build, and
one that just checks the environment and executes when needed.
Also make sure the staging/ directory has the right permissions.