Improve validation of array inputs.
jb::fftw::time_delay_estimator was validating the
number of elements in the input arrays to estimate().
It needed to compare the shapes, because just matching
the element count can succeed on N*M vs M*N arrays.
Also improved the testing to compare more branches,
but that won't come up in the code (line) coverage.