Fixed AM_CXXFLAGS vs. AM_CPPFLAGS vs. CXXFLAGS.
This stuff is not enjoyable.  I made a mistake in previous
changes and did not pass the AM_CXXFLAGS settings from
configure.ac to the Makefile.  Fixing that exposed another
problem: I was putting preprocessor settings in AM_CXXFLAGS.
Fixed that, and then discovered that *some* preprocessor
settings (notably -DBOOST_LOG_DYN_LINK) **must** be in the
CXXFLAGS, grrrrr...  All good for my builds, let's see
the rest of the platforms.