Add support to benchmark jb::opencl::generic_reduce.
This change introduces support to benchmark generic_reduce,
it required changes to the interface in jb::opencl::generic_reduce
because the benchmark needs to call the same reducer with
multiple different sizes.  That, in turn, required changes to
the unit tests.  I also discovered problems with OpenCL
implementations that are more strict about double precision
extensions.