Fix test on hosts with only IPv4 support.
The travis-ci.org build hosts do not seem to have IPv6 enabled.
We need to figure out, at runtime, if the host supports IPv6
when validating the creation of UDP channels.  Ideally one
would make the tests hermetic by using a mock UDP socket,
but that is a lot of work and we need to test IPv4 anyway.