More flakiness fixes, the tolerance was too low.
The test is comparing a numerical result against expected values,
depending on the platform the results might be off by a few units,
this is expected, the test tolerance was simply too tight.
I do not have a good way to know what the correct tolerance should
be from first principles, so adjust the test for the time being.