Fix race condition in completion_queue.
There was a difficult race condition on the completion queue.
Basically I wanted to store the result of calling an asynchronous
operation in the async_op holder, but in an attempt of being
efficient the code was moving the holder during the callback.
If the callback is called before the asynchronous operation returns
it can be destroy the location where the result will be stored,
oops.

Notice that the async_op itself is safe, by design their lifetime
exceeds the lifetime of the callback.  The problem was a member
variable inside the async_op.